apps/backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.9.2
pydantic-settings==2.1.0
slowapi==0.1.9
psutil==5.9.8

# Monitoring & Error Tracking
sentry-sdk[fastapi]==1.40.0

# Database
motor==3.3.2
redis==5.0.1
pymongo==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
authlib==1.3.0
itsdangerous==2.1.2

# File Upload & Processing
python-multipart==0.0.6
cloudinary==1.38.0
pillow==11.0.0

# WebSocket
python-socketio==5.11.0
websockets==12.0

# Task Queue
redis==5.0.1
rq==1.16.1

# Email/SMS
aiosmtplib==3.0.1
emails==0.6

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
geopy==2.4.1

# Development
pytest==8.0.0
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.15
